shortener cleanup
