+++ internal/shortener/shortener.go
@@ -1,9 +1,34 @@
 package shortener
 
-import "encoding/base64"
+import (
+	"encoding/base64"
+	"net/url"
+	"os"
+)
+
+const (
+	BaseURLEnvName = "BASE_URL"
+	DefaultBaseURL = "http://localhost:8080"
+)
+
+var baseURL url.URL
+
+func init() {
+	baseURL = parseBaseURL()
+}
+
+func parseBaseURL() url.URL {
+	baseURL, err := url.Parse(os.Getenv(BaseURLEnvName))
+	if err != nil || (*baseURL == url.URL{}) {
+		baseURL, _ = url.Parse(DefaultBaseURL)
+	}
+	return *baseURL
+}
 
 func Encode(v []byte) string {
-	return base64.StdEncoding.EncodeToString(v)
+	url := baseURL
+	url.Path = base64.StdEncoding.EncodeToString(v)
+	return url.String()
 }
 
 func Decode(v string) ([]byte, error) {
